All -> Refactor Hero section to use new SectionTitle component and enhance Title animations
